% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes-traits-functions.R
\name{read.stones}
\alias{read.stones}
\title{Read BayesTraits Stones files}
\usage{
read.stones(filename)
}
\arguments{
\item{filename}{the path to the BayesTraits Stones file}
}
\value{
the marginal likelihood
}
\description{
Given a the path of a BayesTraits Stones file, this function will retrieve the marginal likelihood from the bottom of that file.
}
\keyword{BayesTraits,}
\keyword{excd,}
\keyword{phylogeny}
