% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treedata2.R
\name{treedata2}
\alias{treedata2}
\title{Compare taxa in dataframe and phylogeny}
\usage{
treedata2(phy, data, sort = FALSE, warnings = TRUE)
}
\description{
This is the function treedata from the geiger package, however it has been altered to output a dataframe instead of a matrix when given a dataframe. See ?geiger::treedata for more help materials
}
\keyword{excd,}
\keyword{geiger}
\keyword{phylogeny,}
