% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who-is.R
\name{who_are}
\alias{who_are}
\title{Who is excd?}
\usage{
who_are()
}
\description{
This function takes you to excd.org to let you explore all the great research they are doing!
}
\examples{
excdr::who_are()
}
\keyword{excd}
