% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes-traits-functions.R
\name{write.bayestraits}
\alias{write.bayestraits}
\title{Writing BayesTraits files}
\usage{
write.bayestraits(tree, data, variables, dir = "./", na.omit = FALSE,
  filename)
}
\arguments{
\item{tree}{an object of class phylo or list of phylo objects}

\item{data}{a data.frame}

\item{variables}{a character string or character vector of the names of columns to subset to, alternatively you can use the numeric values to indicate the column}

\item{dir}{the directory to save the output files to. Defaults to current directory}

\item{na.omit}{a TRUE/FALSE argument to specify whether NA values should be removed from your final data file}

\item{optional:}{used to name the output file. Will use variable names by default.}
}
\value{
two files will be saved to the specified directory *.bttrees holding the pruned tree files and *.btdata holding the subset and formatted data.
}
\description{
Given a data.frame and either a single object of class phylo or a list of class phylo objects, this function will subset the data on specified variables and prune the trees to match that data (optionally removing NA values) and write the data into a format suitable for use with BayesTraits.
}
\keyword{BayesTraits,}
\keyword{excd,}
\keyword{phylogeny}
